Keep PDF render mapping with extractor
